% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalTemp-package.R
\docType{package}
\name{globalTemp-package}
\alias{globalTemp}
\alias{globalTemp-package}
\title{globalTemp: Data Package with Global Temperature Change}
\description{
The package provides monthly global temperature change data, from 1880 to 2022. The data comes from the NASA website.
}
\author{
\strong{Maintainer}: Alexandre Bohyn \email{alexandre.bohyn@gmail.com} (\href{https://orcid.org/0000-0001-9776-7467}{ORCID})

}
\keyword{internal}
